-[BACKEND] Upgrade the references to the latest DataManagement 0.19 library. Some classes have changed name and moved to another package.
Former-commit-id: e3b35c0a37da71f9741aaf610b295ab79a126421

+++ NeoCom.Infinity.Backend/backend.AcceptanceTest/src/test/java/org/dimensinfin/eveonline/neocom/infinity/steps/NIB04Contracts.java
@@ -61,7 +61,7 @@ public class NIB04Contracts extends SupportSteps {
 				Assert.assertEquals( row.get( JWT_TOKEN ), responseValidation.getJwtToken() );
 				Assert.assertEquals( row.get( CREDENTIAL ), responseValidation.getCredential().getJsonClass() );
 				Assert.assertEquals( Integer.valueOf( row.get( ACCOUNT_ID ) ).intValue(),
-						responseValidation.getCredential().getAccountId() );
+						responseValidation.getCredential().getAccountId().intValue() );
 				Assert.assertEquals( Integer.valueOf( row.get( CORPORATION_ID ) ).intValue(),
 						responseValidation.getCredential().getCorporationId() );
 				Assert.assertEquals( row.get( ACCOUNT_NAME ), responseValidation.getCredential().getAccountName() );
